完成了大部分后端开发，明天测试接口，适当修改后开始写前端

+++ src/main/java/lds/com/medicalsystem/staff/admin/controller/AdminArrangeController.java
@@ -1,13 +1,13 @@
 package lds.com.medicalsystem.staff.admin.controller;
 
 import lds.com.medicalsystem.common.VO.ResultVO;
+import lds.com.medicalsystem.common.utils.exception.BusinessException;
 import lds.com.medicalsystem.staff.admin.DTO.DocOtherInfoDTO;
 import lds.com.medicalsystem.staff.admin.DTO.SearchTableDTO;
 import lds.com.medicalsystem.staff.admin.DTO.WorkTableInsertDTO;
 import lds.com.medicalsystem.staff.admin.DTO.WorkTableUpdateDTO;
 import lds.com.medicalsystem.staff.admin.VO.PageResultVO;
-import lds.com.medicalsystem.staff.admin.VO.WorkListVO;
-import lds.com.medicalsystem.staff.admin.VerifyUtil;
+import lds.com.medicalsystem.staff.VerifyUtil;
 import lds.com.medicalsystem.staff.admin.service.AdminArrangeService;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,53 +24,60 @@ public class AdminArrangeController {
     //返回某个科室分类下的医生名列表
     @GetMapping("/docList/deptSort")
     public ResultVO<List<String>> DocNameListInDeptSort(@RequestParam String deptSortName){
-        VerifyUtil.adminVerify();
-        List<String> docNameList = adminArrangeService.docNameListInDeptSort(deptSortName);
-        return ResultVO.success(docNameList);
+        if(VerifyUtil.adminVerify()) {
+            List<String> docNameList = adminArrangeService.docNameListInDeptSort(deptSortName);
+            return ResultVO.success(docNameList);
+        } else throw new BusinessException("非法角色");
     }
     //返回某个科室下的医生名列表
     @GetMapping("/docList/dept")
     public ResultVO<List<String>> DocNameListInDept(@RequestParam String deptName){
-        VerifyUtil.adminVerify();
-        List<String> docNameList = adminArrangeService.docNameListInDept(deptName);
-        return ResultVO.success(docNameList);
+        if(VerifyUtil.adminVerify()) {
+            List<String> docNameList = adminArrangeService.docNameListInDept(deptName);
+            return ResultVO.success(docNameList);
+        } else throw new BusinessException("非法角色");
     }
     //根据医生名查工号、科室（修改、添加时选择医生名称，工号、科室自动填充）
     @GetMapping("/findNoAndDept")
     public ResultVO<DocOtherInfoDTO> searchOtherByDocName(@RequestParam String docName) {
-        VerifyUtil.adminVerify();
-        DocOtherInfoDTO docOtherInfoDTO = adminArrangeService.searchOtherByDocName(docName);
-        return ResultVO.success(docOtherInfoDTO);
+        if(VerifyUtil.adminVerify()) {
+            DocOtherInfoDTO docOtherInfoDTO = adminArrangeService.searchOtherByDocName(docName);
+            return ResultVO.success(docOtherInfoDTO);
+        } else throw new BusinessException("非法角色");
     }
 
     //设置动态sql，点击值班管理，默认显示全部科室(全部分类下的全部科室），当只选择科室分类的时候显示该科室分类下所有科室对应的对象列表
     // 当选择了具体科室后显示该具体科室下所有科室对应的对象列表，且都具有分页功能，前端传入页码、pageSize、deptSort、deptName
     @PostMapping("/workList/search")
     public ResultVO<PageResultVO> searchWorkList(@RequestBody SearchTableDTO dto){
-        VerifyUtil.adminVerify();
-        PageResultVO pageResultVO = adminArrangeService.searchWorkList(dto);
-        return ResultVO.success(pageResultVO);
+        if(VerifyUtil.adminVerify()) {
+            PageResultVO pageResultVO = adminArrangeService.searchWorkList(dto);
+            return ResultVO.success(pageResultVO);
+        } else return ResultVO.error("非法角色");
     }
     //管理员选择值班日期、时间段、值班医生工号(触发自动匹配医生姓名、所属科室）、最大预约人数，而后插入表
     @PostMapping("/workList/insert")
     public ResultVO<Void> insertWorkInfo(WorkTableInsertDTO dto){
-        VerifyUtil.adminVerify();
-        adminArrangeService.insertWorkInfo(dto);
-        return ResultVO.success("插入值班信息成功");
+        if(VerifyUtil.adminVerify()) {
+            adminArrangeService.insertWorkInfo(dto);
+            return ResultVO.success("插入值班信息成功");
+        } return ResultVO.error("非法角色");
     }
     //点击列表项的编辑图标可查看单项详细信息（前端通过列表单项信息渲染，非后端工作）
     //可编辑修改单项内容，点击保存修改至数据库
     @PatchMapping("/workList/update")
     public ResultVO<Void> updateWorkInfo(WorkTableUpdateDTO dto){
-        VerifyUtil.adminVerify();
-        adminArrangeService.updateWorkInfo(dto);
-        return ResultVO.success("修改值班信息成功");
+        if(VerifyUtil.adminVerify()) {
+            adminArrangeService.updateWorkInfo(dto);
+            return ResultVO.success("修改值班信息成功");
+        } return ResultVO.error("非法角色");
     }
     //可删除单项内容
     @DeleteMapping("/workList/delete")
     public ResultVO<Void> deleteWorkInfo(Integer id){
-        VerifyUtil.adminVerify();
-        adminArrangeService.deleteWorkInfo(id);
-        return ResultVO.success("删除值班信息成功");
+        if(VerifyUtil.adminVerify()) {
+            adminArrangeService.deleteWorkInfo(id);
+            return ResultVO.success("删除值班信息成功");
+        } return ResultVO.error("非法角色");
     }
 }
